# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Mother Control Plane
  version: "0.1.0"
  description: Central API server for Mother. Dispatches requests to services.

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/v1/jobs:
    post:
      operationId: createJob
      summary: Submit a new job
      description: |
        Creates a new async job. Currently only supports the "coder" service.
        Returns immediately with a job ID. Poll GET /api/v1/jobs/{id} for status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/jobs/{id}:
    get:
      operationId: getJob
      summary: Get job status and result
      description: Retrieves the current status, and result if complete, for a job.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      description: Returns server health status.
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

components:
  schemas:
    CreateJobRequest:
      type: object
      required:
        - service
        - params
      properties:
        service:
          type: string
          enum:
            - coder
          description: The service to invoke.
        params:
          $ref: "#/components/schemas/CoderParams"

    CoderParams:
      type: object
      required:
        - project_dir
        - prompt
      properties:
        project_dir:
          type: string
          description: Absolute path to the project directory.
        prompt:
          type: string
          description: The prompt to send to Claude Code.
        env_vars:
          type: object
          additionalProperties:
            type: string
          description: Environment variables to pass to the coder VM.
        system_prompt:
          type: string
          description: System prompt override for Claude Code.
        model:
          type: string
          description: Model override (e.g., sonnet, opus).

    CreateJobResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique job identifier.

    Job:
      type: object
      required:
        - id
        - service
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        service:
          type: string
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
        params:
          $ref: "#/components/schemas/CoderParams"
        result:
          type: string
          description: Output from the job, available when status is completed.
        error:
          type: string
          description: Error message, available when status is failed.
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.